# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  class.ClassBatchRequest:
    properties:
      academic_year_id:
        type: string
      capacity:
        type: integer
      count:
        minimum: 1
        type: integer
      level_ids:
        items:
          type: string
        minItems: 1
        type: array
      major_code:
        description: 'Untuk digabung dalam nama, ex: IPA'
        type: string
      major_id:
        type: string
    required:
    - academic_year_id
    - count
    - level_ids
    - major_code
    - major_id
    type: object
  class.ClassRequest:
    properties:
      academic_year_id:
        type: string
      capacity:
        type: integer
      class_name:
        type: string
      classroom:
        type: string
      homeroom_teacher_id:
        type: string
      level_id:
        type: string
      major_id:
        type: string
    required:
    - academic_year_id
    - class_name
    - level_id
    - major_id
    type: object
  class.ClassResponse:
    properties:
      academic_year_id:
        type: string
      capacity:
        type: integer
      class_name:
        type: string
      classroom:
        type: string
      created_at:
        type: string
      homeroom_teacher_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      level_id:
        type: string
      major_id:
        type: string
      updated_at:
        type: string
    type: object
  class.ClassUpdateRequest:
    properties:
      capacity:
        type: integer
      homeroom_teacher_id:
        type: string
    type: object
  major.MajorRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  major.MajorResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  response.DefaultResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Documentation for LinguaLoop E-Learning Platform.
  title: LinguaLoop API
  version: "1.0"
paths:
...
  /classes:
    get:
      description: Get a list of all classes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/class.ClassResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all classes
      tags:
      - Classes
    post:
      consumes:
      - application/json
      description: Add a single new class to the system
      parameters:
      - description: Class Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.ClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/class.ClassResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create new class
      tags:
      - Classes
  /classes/{id}:
    delete:
      description: Soft delete a class by ID
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete class
      tags:
      - Classes
    get:
      description: Get detailed information of a class by its ID
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/class.ClassResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get class by ID
      tags:
      - Classes
    put:
      consumes:
      - application/json
      description: Update existing class capacity or homeroom teacher
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      - description: Class Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.ClassUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/class.ClassResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update class
      tags:
      - Classes
  /classes/batch:
    post:
      consumes:
      - application/json
      description: Auto-generate multiple classes alphabetically per level based on
        existing DB records
      parameters:
      - description: Batch Class Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.ClassBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/class.ClassResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create classes in batch
      tags:
      - Classes
  /majors:
    get:
      description: Retrieve a list of all majors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/major.MajorResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get All Majors
      tags:
      - Majors
    post:
      consumes:
      - application/json
      description: Create a new major
      parameters:
      - description: Major Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/major.MajorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/major.MajorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create Major
      tags:
      - Majors
  /majors/{id}:
    delete:
      description: Delete a major by ID
      parameters:
      - description: Major ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete Major
      tags:
      - Majors
    put:
      consumes:
      - application/json
      description: Update an existing major by ID
      parameters:
      - description: Major ID
        in: path
        name: id
        required: true
        type: string
      - description: Major Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/major.MajorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/major.MajorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update Major
      tags:
      - Majors
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"